move endpoint test in NodeTasksTest to fusion-endpoint
