+++ flow-server/src/test/java/com/vaadin/flow/server/frontend/NodeTasksTest.java
@@ -4,13 +4,10 @@ import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.Field;
 import java.nio.file.Files;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.stream.Collectors;
 
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -18,7 +15,6 @@ import org.mockito.Mockito;
 
 import com.vaadin.flow.di.Lookup;
 import com.vaadin.flow.server.ExecutionFailedException;
-// import com.vaadin.flow.server.connect.Endpoint;
 import com.vaadin.flow.server.frontend.NodeTasks.Builder;
 import com.vaadin.flow.server.frontend.scanner.ClassFinder.DefaultClassFinder;
 
@@ -30,13 +26,8 @@ import static com.vaadin.flow.server.frontend.FrontendUtils.PARAM_GENERATED_DIR;
 import static com.vaadin.flow.server.frontend.FrontendUtils.TARGET;
 import static com.vaadin.flow.server.frontend.FrontendUtils.WEBPACK_CONFIG;
 import static com.vaadin.flow.server.frontend.FrontendUtils.WEBPACK_GENERATED;
-import static org.junit.Assert.assertTrue;
 
 public class NodeTasksTest {
-
-    // @Endpoint
-    public static class ConnectEndpointsForTesting {
-    }
 
     @Rule
     public TemporaryFolder temporaryFolder = new TemporaryFolder();
@@ -113,37 +104,6 @@ public class NodeTasksTest {
                 "useClientSideIndexFileForBootstrapping should be true",
                 webpackGeneratedContent.contains(
                         "const useClientSideIndexFileForBootstrapping = true;"));
-    }
-
-    @Test
-    @Ignore
-    public void should_Generate_Connect_Files() throws Exception {
-        File src = new File(getClass().getClassLoader().getResource("java").getFile());
-        File dir = new File(userDir);
-        File json = new File(dir, "api-file.json");
-
-        Builder builder = new Builder(
-                new DefaultClassFinder(
-                    Collections.singleton(ConnectEndpointsForTesting.class)), dir, Mockito.mock(Lookup.class))
-                        .enablePackagesUpdate(false)
-                        .enableImportsUpdate(false)
-                        .withEmbeddableWebComponents(false)
-                        .withConnectJavaSourceFolder(src)
-                        .withConnectGeneratedOpenApiJson(json)
-                        .withConnectClientTsApiFolder(new File(dir, "api"));
-
-        builder.build().execute();
-
-        Arrays.asList(
-                // enableClientSide
-                "target/index.html",
-                "target/index.ts",
-                // withConnectJavaSourceFolder and withConnectGeneratedOpenApiJson
-                "api-file.json",
-                // withConnectClientTsApiFolder
-                "api/connect-client.default.ts",
-                "api/MyEndpoint.ts")
-                .forEach(name -> assertTrue(name + " not created.", new File(dir, name).exists()));
     }
 
     private Object getFieldValue(Object obj, String name) throws Exception {
